Move test setup to be near tests
