+++ spec/graphql/schema_spec.rb
@@ -220,6 +220,48 @@ describe GraphQL::Schema do
   end
 
   describe "`use` works with plugins that attach instrumentation, tracers, query analyzers" do
+    class NoOpTracer
+      def trace(_key, data)
+        if (query = data[:query])
+          query.context[:no_op_tracer_ran] = true
+        end
+        yield
+      end
+    end
+
+    class NoOpInstrumentation
+      def before_query(query)
+        query.context[:no_op_instrumentation_ran_before_query] = true
+      end
+
+      def after_query(query)
+        query.context[:no_op_instrumentation_ran_after_query] = true
+      end
+    end
+
+    class NoOpAnalyzer < GraphQL::Analysis::AST::Analyzer
+      def initialize(query_or_multiplex)
+        query_or_multiplex.context[:no_op_analyzer_ran_initialize] = true
+        super
+      end
+
+      def on_leave_field(_node, _parent, visitor)
+        visitor.query.context[:no_op_analyzer_ran_on_leave_field] = true
+      end
+
+      def result
+        query.context[:no_op_analyzer_ran_result] = true
+      end
+    end
+
+    module PluginWithInstrumentationTracingAndAnalyzer
+      def self.use(schema_defn)
+        schema_defn.instrument :query, NoOpInstrumentation.new
+        schema_defn.tracer NoOpTracer.new
+        schema_defn.query_analyzer NoOpAnalyzer
+      end
+    end
+
     query_type = Class.new(GraphQL::Schema::Object) do
       graphql_name 'Query'
       field :foobar, Integer, null: false
